requirements.txt
# Le coeur du système
llama-index-core

# Les nouvelles librairies Google (celles qui marchent)
llama-index-llms-google-genai
llama-index-embeddings-google-genai

# La base de données
llama-index-vector-stores-qdrant
qdrant-client

# Outils
python-dotenv
pypdf
fastapi
uvicorn